finally fixed it all. wrote a simple backend to handle everything
