+++ src/component/display/display.tsx
@@ -7,22 +7,6 @@ const Card = (project: Md) => {
 
 //const Display = () => {
 const Display = ({ projects = [] }: any) => {
-
-  /**
-  if(projects.length == 0){
-    const m: Md = {
-      title: "dummy_title",
-      description: "dummy_description",
-      link: "dummy_link",
-    }
-    projects = [
-      m
-    ]
-  }
-  **/
-
-  console.log("loaded projects", projects);
-
   return (
       <div className={styles['column']}>
         {
